+++ .phpunit.result.cache
@@ -1,0 +1,1 @@
+{"version":"pest_2.36.0","defects":{"P\\Tests\\Unit\\TurnstileServiceProviderTest::__pest_evaluable_it_test":8},"times":{"P\\Tests\\Unit\\TurnstileServiceProviderTest::__pest_evaluable_it_test":0.013}}
