+++ config.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+  'secret_key' =>  'key',
+];
