+++ src/config/turnstile.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    // Default turnstile key
+    'key' => env('TURNSTILE_KEY', 'your-default-key'),
+];
